# docs/swagger.yaml
basePath: /v1
definitions:
  httpx.DataJsonResponse:
    properties:
      data: {}
    type: object
  httpx.ErrorJsonResponse:
    properties:
      error:
        type: string
    type: object
  httpx.MessageJsonResponse:
    properties:
      message:
        type: string
    type: object
  schemas.AddStudentRequest:
    properties:
      email:
        type: string
      fullname:
        type: string
      nisn:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
...
  schemas.DeleteStudentRequest:
    properties:
      id:
        type: string
    type: object
...
  schemas.Student:
    properties:
      email:
        type: string
      fullname:
        type: string
      id:
        type: string
      nisn:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
  schemas.StudentProfile:
    properties:
      email:
        type: string
      fullname:
        type: string
      id:
        type: string
      nisn:
        type: string
      phone_number:
        type: string
      school:
        type: string
      username:
        type: string
    type: object
...
  schemas.UpdateStudentRequest:
    properties:
      email:
        type: string
      fullname:
        type: string
      nisn:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: This is a Double Track REST API
  title: Double Track API
  version: "1.0"
paths:
...
  /students:
    delete:
      consumes:
      - application/json
      description: Delete a student from a school
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Delete student request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.DeleteStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Delete a student
      tags:
      - Students
      - Admin
    get:
      consumes:
      - application/json
      description: Get all students from a school
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schemas.Student'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get students
      tags:
      - Students
      - Admin
    post:
      consumes:
      - application/json
      description: Add a student from a school
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add student request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.AddStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Add a student
      tags:
      - Students
      - Admin
  /students/{studentId}:
    get:
      consumes:
      - application/json
      description: Get a student from a school
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Id student
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.Student'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get a student
      tags:
      - Students
      - Admin
    patch:
      consumes:
      - application/json
      description: Update a student from a school
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update student id
        in: path
        name: studentId
        required: true
        type: string
      - description: Update student request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Update a student
      tags:
      - Students
      - Admin
  /students/profile:
    get:
      consumes:
      - application/json
      description: Get a student profile
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.StudentProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get a student profile
      tags:
      - Students
      - Student
    patch:
      consumes:
      - application/json
      description: Update a student profile
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update student profile body request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Update a student profile
      tags:
      - Students
      - Student
swagger: "2.0"